+++ src/analytics/process_pattern_matcher.cpp
@@ -117,8 +117,8 @@ std::vector<float> ProcessPatternMatcher::embedActivities(const std::vector<std:
         // Pad with sentinel chars
         std::string s = " " + act + " ";
         for (size_t i = 0; i + 2 <static_cast<int>(s.size()); ++i) {
-            uint32_t h = (static_cast<uint32_t>(static_cast<unsigned char>(s[i])) * 31u * 31u
-                          + static_cast<uint32_t>(static_cast<unsigned char>(s[i + 1])) * 31u
+            uint32_t h = (static_cast<uint32_t>(static_cast<unsigned char>(s[i])) * 31 * 31
+                          + static_cast<uint32_t>(static_cast<unsigned char>(s[i + 1])) * 31
                           + static_cast<uint32_t>(static_cast<unsigned char>(s[i + 2])))
                          % static_cast<uint32_t>(DIM);
             vec[h] += 1.0;
@@ -178,8 +178,8 @@ double ProcessPatternMatcher::computeGraphSimilarity(const ProcessPattern &patte
     std::set<std::string> intersection_set;
     std::set_intersection(pat_set.begin(), pat_set.end(), trace_set.begin(), trace_set.end(),
                           std::inserter(intersection_set, intersection_set.begin()));
-    size_t sym_diff  = (static_cast<int>(pat_set.size()) + trace_set.size()) - 2 * intersection_set.size();
-    double denom     = static_cast<double>(static_cast<int>(pat_set.size()) + trace_set.size());
+    size_t sym_diff  = (static_cast<int>(pat_set.size()) + static_cast<int>(trace_set.size()) ) - 2 * intersection_set.size();
+    double denom     = static_cast<double>(static_cast<int>(pat_set.size()) + static_cast<int>(trace_set.size()) );
     double edit_norm = (denom > 0) ? 1.0 - static_cast<double>(sym_diff) / denom : 1.0;
 
     return 0.30 * node_jac + 0.30 * edge_jac + 0.25 * path_sim + 0.15 * edit_norm;
@@ -373,8 +373,8 @@ ProcessPatternMatcher::findSimilar(const ProcessPattern &pattern, const PatternM
         std::set<std::string> inter_set;
         std::set_intersection(pat_set.begin(), pat_set.end(), trace_set.begin(), trace_set.end(),
                               std::inserter(inter_set, inter_set.begin()));
-        size_t sym_diff       = (static_cast<int>(pat_set.size()) + trace_set.size()) - 2 * inter_set.size();
-        double den            = static_cast<double>(static_cast<int>(pat_set.size()) + trace_set.size());
+        size_t sym_diff       = (static_cast<int>(pat_set.size()) + static_cast<int>(trace_set.size()) ) - 2 * inter_set.size();
+        double den            = static_cast<double>(static_cast<int>(pat_set.size()) + static_cast<int>(trace_set.size()) );
         metrics.edit_distance = (den > 0) ? static_cast<double>(sym_diff) / den : 0.0;
 
         // Matched / missing / extra activities
