+++ src/aql/aql_query_diff_explainer.cpp
@@ -83,7 +83,7 @@ std::unordered_map<std::string, std::string> splitClauses(const std::string &nor
                 break;
             }
             bool ok_before = (p == 0 || !std::isalnum(static_cast<unsigned char>(norm[static_cast<int>(p - 1)])));
-            size_t after   = p + kw.size();
+            size_t after   = p + static_cast<int>(kw.size()) ;
             bool ok_after  = (after >= norm.size() || !std::isalnum(static_cast<unsigned char>(norm[after])));
             if (ok_before && ok_after) {
                 matches.push_back({p, kw});
